🎨 Change type param to only float

+++ src/services/cotization.py
@@ -19,7 +19,7 @@ class Cotization:
             data = response.json()
             return data["RAW"]["USDT"]["ARS"]["PRICE"]
 
-    async def calculate_cotization(self, buy_price: int) -> float:
+    async def calculate_cotization(self, buy_price: float) -> float:
         cotization_value: float | None = await self._obtain()
         cotization_market_int = int(cotization_value or 0)
         total_cotization: int | float = round((buy_price / cotization_market_int), 2)
